v0.9.8
Updated order product weight type from Int to Float.

+++ server/graphql/types/order-product/order-product-patch.ts
@@ -19,7 +19,7 @@ export const OrderProductPatch = gql`
     batchId: String
     packingType: String
     unit: String
-    weight: Int
+    weight: Float
     packQty: Int
     actualPackQty: Int
     palletQty: Int
